Derive signal name/number from one authoritative SIGNALS table
The signal-menu code encoded the same 9-signal mapping three times: the
SIGNALS table (label + number + Signal), signal_name() (Signal -> label),
and raw_signal() (Signal -> number). Adding a signal meant editing three
lists that could silently drift.

Move SIGNALS into the metrics layer (its natural owner, and it lets
raw_signal derive from it without an upward metrics -> app dependency) and
derive both helpers by lookup over the Signal key. signal_name and
raw_signal are now one-liners; a new signal is a single-row edit.

Also name the battery poll interval (BATTERY_POLL) instead of a bare 10.

No behavior change; fmt/clippy clean, tests green.
